:art: chore: 重构 const token
