+++ src/constants/common.ts
@@ -2,37 +2,8 @@ export const AGENT_INDEX_URL = 'https://vidol-market.lobehub.com/agents/index.js
 
 export const DANCE_INDEX_URL = 'https://vidol-market.lobehub.com/dances/index.json';
 
-export const VIDOL_THEME_APPEARANCE = 'VIDOL_THEME_APPEARANCE';
-export const VIDOL_THEME_NEUTRAL_COLOR = 'VIDOL_THEME_NEUTRAL_COLOR';
-export const VIDOL_THEME_PRIMARY_COLOR = 'VIDOL_THEME_PRIMARY_COLOR';
 export const COOKIE_CACHE_DAYS = 30;
 
 export const LOADING_FLAG = '...';
 
-// 默认坐标
-export const INITIAL_COORDINATES = { x: 360, y: 360 };
-
-export const DESKTOP_HEADER_ICON_SIZE = { fontSize: 24 };
-
-// 默认 zIndex
-export const INITIAL_Z_INDEX = 10;
-
-export const CHAT_TEXTAREA_MAX_HEIGHT = 570;
-export const CHAT_TEXTAREA_HEIGHT = 200;
-
-export const CHAT_INPUT_MIN_HEIGHT = 90;
-
-export const HEADER_HEIGHT = 64;
-
-export const SIDEBAR_WIDTH = 280;
-export const SIDEBAR_MAX_WIDTH = 400;
-
-export const CHAT_HEADER_HEIGHT = 64;
-
-export const CHAT_INPUT_WIDTH = '48rem';
-
 export const DEFAULT_USER_AVATAR = '😀';
-
-export const LIST_GRID_WIDTH = 108;
-export const LIST_GRID_GAP = 4;
-export const LIST_GRID_HEIGHT = 108;
