+++ src/constants/theme.ts
@@ -1,0 +1,3 @@
+export const VIDOL_THEME_APPEARANCE = 'VIDOL_THEME_APPEARANCE';
+export const VIDOL_THEME_NEUTRAL_COLOR = 'VIDOL_THEME_NEUTRAL_COLOR';
+export const VIDOL_THEME_PRIMARY_COLOR = 'VIDOL_THEME_PRIMARY_COLOR';
