+++ src/constants/token.ts
@@ -1,0 +1,16 @@
+// 默认坐标
+export const INITIAL_COORDINATES = { x: 360, y: 360 };
+export const DESKTOP_HEADER_ICON_SIZE = { fontSize: 24 };
+// 默认 zIndex
+export const INITIAL_Z_INDEX = 10;
+export const CHAT_TEXTAREA_MAX_HEIGHT = 570;
+export const CHAT_TEXTAREA_HEIGHT = 200;
+export const CHAT_INPUT_MIN_HEIGHT = 90;
+export const HEADER_HEIGHT = 64;
+export const SIDEBAR_WIDTH = 280;
+export const SIDEBAR_MAX_WIDTH = 400;
+export const CHAT_HEADER_HEIGHT = 64;
+export const CHAT_INPUT_WIDTH = '48rem';
+export const LIST_GRID_WIDTH = 108;
+export const LIST_GRID_GAP = 4;
+export const LIST_GRID_HEIGHT = 108;
