+++ src/constants/touch.ts
@@ -1,0 +1,10 @@
+import { TouchActionConfig } from '@/types/touch';
+
+export const DEFAULT_TOUCH_CONFIG: TouchActionConfig = {
+  arm: [],
+  belly: [],
+  chest: [],
+  enabled: false,
+  head: [],
+  leg: [],
+};
