+++ src/constants/tts.ts
@@ -1,0 +1,9 @@
+import { TTS } from '@/types/tts';
+
+export const DEFAULT_TTS_CONFIG: TTS = {
+  engine: 'edge',
+  locale: 'zh-CN',
+  pitch: 1,
+  speed: 1,
+  voice: 'zh-CN-XiaoxiaoNeural',
+};
